Fix crash: don't create py::bytes inside request callback
Requests were crashing and returning strange results because the GIL
wasn't held in the callback that was creating python `bytes` instances
in the callback.  Actually acquiring the gil, however, deadlocks.

So instead this keeps the future result as a C++ type and converts it to
python types when `get()` is called from Python.